modifying user social media

+++ app/graphql/UserSocialNetwork/UserSocialNetowrkResolver.py
@@ -5,18 +5,8 @@ from app.config.logger import logger
 from sqlalchemy.exc import IntegrityError, SQLAlchemyError
 from strawberry.exceptions import GraphQLError
 from app.utils.helpers import Helpers
-
-@strawberry.input
-class SocialMediaStoreInput:
-    socialMediaId: int
-    link: str
-
-@strawberry.type
-class SocialMediPayload:
-    userSocialNetworkId: int
-    socialMediaId: int
-    network: str
-    link: str
+from app.graphql.UserSocialNetwork.UserSocialNetworkInputs import SocialMediaStoreInput, UpdateUserNetworkInput
+from app.graphql.UserSocialNetwork.UserSocialNetworkPayloads import SocialMediPayload
 
 @strawberry.type
 class UserSocialNetworkQuery:
@@ -100,7 +90,7 @@ class UserSocialNetworkMutation:
             raise GraphQLError(message=error_message, extensions={"code": extension_code})
         
     @strawberry.mutation
-    async def removeUserSocialNetwork(self, info, userSocialNetworkId: int) -> str:
+    async def updateUserSocialNetwork(self, info, updateUserNetwork: UpdateUserNetworkInput) -> str:
         db = info.context["db"]
         current_user = info.context["current_user"]
         user_service = UserService(db)
@@ -110,8 +100,60 @@ class UserSocialNetworkMutation:
             if not user.get("ok", False):
                 raise GraphQLError(message=user['error'], extensions={"code": "BAD_USER_INPUT"})
 
+            item = await usr_scl_ntw_service.getUserSocialMediaById(updateUserNetwork.userSocialNetworkId)
+            if not item.get("ok", False):
+                raise GraphQLError(message=item['error'], extensions={"code": "BAD_USER_INPUT"})
+            
+            item = item.get("data")
+
+            update = await usr_scl_ntw_service.update(
+                item,
+                updateUserNetwork.socialMediaId,
+                updateUserNetwork.link
+            )
+
+            if update.get("ok", False):
+                return update.get("message")
+            
+            raise GraphQLError(message=update['error'], extensions={"code": "BAD_USER_INPUT"})
+
+        except GraphQLError as e:
+            logger.error(e.message)
+            raise e
+
+        except Exception as e:
+            error_mapping = {
+                IntegrityError: ("BAD_USER_INPUT", "El correo ya está en uso"),
+                SQLAlchemyError: ("INTERNAL_SERVER_ERROR", "Error interno del servidor"),
+                ValueError: ("BAD_USER_INPUT", "Datos inválidos"),
+                PermissionError: ("FORBIDDEN", "Permiso denegado"),
+                FileNotFoundError: ("NOT_FOUND", "Archivo no encontrado"),
+                ConnectionError: ("TOO_MANY_REQUESTS", "Demasiadas solicitudes"),
+            }
+
+            extension_code, error_message = error_mapping.get(type(e), ("INTERNAL_SERVER_ERROR", "Error desconocido"))
+            logger.error(error_message)
+            raise GraphQLError(message=error_message, extensions={"code": extension_code})
+        
+    @strawberry.mutation
+    async def removeUserSocialNetwork(self, info, userSocialNetworkId: int) -> str:
+        db = info.context["db"]
+        current_user = info.context["current_user"]
+        user_service = UserService(db)
+        usr_scl_ntw_service = UserSocialNetworkService(db)
+        try:
+            user = await user_service.getUserById(current_user.userId)
+            if not user.get("ok", False):
+                raise GraphQLError(message=user['error'], extensions={"code": "BAD_USER_INPUT"})
+            
+            item = await usr_scl_ntw_service.getUserSocialMediaById(userSocialNetworkId)
+            if not item.get("ok", False):
+                raise GraphQLError(message=item['error'], extensions={"code": "BAD_USER_INPUT"})
+            
+            item = item.get("data")
+
             remove = await usr_scl_ntw_service.remove(
-                userSocialNetworkId
+                item
             )
 
             if remove.get("ok", False):
